requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.22

sqlalchemy>=2.0.30,<3
asyncpg>=0.27
psycopg2-binary>=2.9
geoalchemy2>=0.14
alembic>=1.12

python-dotenv>=1.0
requests>=2.31
httpx>=0.24

numpy>=1.26
pandas>=2.2
xarray>=2024.1
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
pyproj>=3.6

scikit-learn>=1.2
joblib>=1.3
apscheduler>=3.10

planetary-computer>=1.0
pystac-client>=0.7.6
pydantic>=2.6,<3
